Move relationship playground to proper namespace
* Relationship playground types are now moved to "Playground"
  sub-folder, separating it from V1.
* Tests edited accordingly
